APIKIT-2822: Revert packages refactoring (#170)
* Revert "APIKIT-2821: Move ParserService to internal package (#167)"

This reverts commit aaea63fdc2e2b54e08390dc421dc256624885fe1.

* Revert "APIKIT-2821: Expose code for identifying the error type (#165)"

This reverts commit 31d040d1c75e505155b0621a4f34125cd6ae0cef.

* Revert "APIKIT-2821: Move result implementations to internal package (#163)"

This reverts commit 2a26a1b519847ca22cbdcf04a67b41650773e445.
