+++ parser-service/src/main/java/org/mule/parser/service/result/ExceptionParseResult.java
@@ -4,17 +4,15 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.parser.service.result.internal;
+package org.mule.parser.service.result;
+
+import static java.util.Collections.*;
 
 import org.mule.apikit.model.ApiSpecification;
-import org.mule.parser.service.result.ParsingIssue;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.List;
-
-import static java.util.Collections.emptyList;
-import static java.util.Collections.singletonList;
 
 public class ExceptionParseResult implements ParseResult {
