+++ parser-service/src/main/java/org/mule/parser/service/result/ExceptionParsingIssue.java
@@ -4,18 +4,24 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.parser.service.result.internal;
+package org.mule.parser.service.result;
 
 /**
  * Represents a simple parsing error with cause message
  */
-public class ExceptionParsingIssue extends DefaultParsingIssue {
+public class ExceptionParsingIssue implements ParsingIssue {
 
+  private final String cause;
   private final String stackTrace;
 
   public ExceptionParsingIssue(String cause, String stackTrace) {
-    super(cause);
+    this.cause = cause;
     this.stackTrace = stackTrace;
+  }
+
+  @Override
+  public String cause() {
+    return cause;
   }
 
   public String getStackTrace() {
@@ -25,7 +31,7 @@ public class ExceptionParsingIssue extends DefaultParsingIssue {
   @Override
   public String toString() {
     return "ExceptionParsingIssue {" +
-        "cause='" + cause() + '\'' +
+        "cause='" + cause + '\'' +
         ", stackTrace='" + stackTrace + '\'' +
         '}';
   }
