+++ parser-service/src/main/java/org/mule/parser/service/result/ParseResult.java
@@ -4,10 +4,9 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.parser.service.result.internal;
+package org.mule.parser.service.result;
 
 import org.mule.apikit.model.ApiSpecification;
-import org.mule.parser.service.result.ParsingIssue;
 
 import java.util.List;
