+++ parser-service/src/main/java/org/mule/parser/service/strategy/ParsingStrategy.java
@@ -6,10 +6,9 @@
  */
 package org.mule.parser.service.strategy;
 
-import org.mule.apikit.model.api.ApiReference;
-import org.mule.parser.service.result.internal.ParseResult;
-
 import java.util.concurrent.ScheduledExecutorService;
+import org.mule.apikit.model.api.ApiReference;
+import org.mule.parser.service.result.ParseResult;
 
 public interface ParsingStrategy {
