+++ parser-service/src/main/java/org/mule/parser/service/strategy/RamlParsingStrategy.java
@@ -6,6 +6,13 @@
  */
 package org.mule.parser.service.strategy;
 
+import static java.util.Collections.singletonList;
+import static org.mule.apikit.model.ApiVendor.RAML_08;
+import static org.mule.parser.service.strategy.ValidationReportHelper.errors;
+import static org.mule.parser.service.strategy.ValidationReportHelper.warnings;
+
+import java.util.List;
+import java.util.concurrent.ScheduledExecutorService;
 import org.mule.apikit.ApiParser;
 import org.mule.apikit.implv1.ParserWrapperV1;
 import org.mule.apikit.implv2.ParserWrapperV2;
@@ -13,17 +20,9 @@ import org.mule.apikit.loader.ResourceLoader;
 import org.mule.apikit.model.api.ApiReference;
 import org.mule.apikit.validation.ApiValidationReport;
 import org.mule.parser.service.references.ReferencesResolver;
-import org.mule.parser.service.result.internal.DefaultParseResult;
-import org.mule.parser.service.result.internal.ExceptionParseResult;
-import org.mule.parser.service.result.internal.ParseResult;
-
-import java.util.List;
-import java.util.concurrent.ScheduledExecutorService;
-
-import static java.util.Collections.singletonList;
-import static org.mule.apikit.model.ApiVendor.RAML_08;
-import static org.mule.parser.service.strategy.ValidationReportHelper.errors;
-import static org.mule.parser.service.strategy.ValidationReportHelper.warnings;
+import org.mule.parser.service.result.DefaultParseResult;
+import org.mule.parser.service.result.ExceptionParseResult;
+import org.mule.parser.service.result.ParseResult;
 
 public class RamlParsingStrategy implements ParsingStrategy {
