+++ parser-service/src/test/java/org/mule/parser/service/ParserServiceTestCase.java
@@ -6,25 +6,24 @@
  */
 package org.mule.parser.service;
 
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.ExpectedException;
-import org.mule.apikit.ApiType;
-import org.mule.apikit.model.api.ApiReference;
-import org.mule.parser.service.internal.ParserService;
-import org.mule.parser.service.result.ParsingIssue;
-import org.mule.parser.service.result.internal.ParseResult;
-
-import java.util.List;
-
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.core.StringContains.containsString;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertThat;
+import static org.junit.Assert.*;
 import static org.mule.apikit.model.ApiVendor.OAS_20;
 import static org.mule.apikit.model.ApiVendor.OAS_30;
 import static org.mule.apikit.model.ApiVendor.RAML_08;
 import static org.mule.apikit.model.ApiVendor.RAML_10;
+
+import org.mule.apikit.ApiType;
+import org.mule.apikit.model.api.ApiReference;
+import org.mule.parser.service.result.ParseResult;
+import org.mule.parser.service.result.ParsingIssue;
+
+import java.util.List;
+
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.ExpectedException;
 
 public class ParserServiceTestCase {
